rename term -> entity file
