+++ src/test/utils/occurrence-test.ts
@@ -1,6 +1,6 @@
 import "mocha";
 import {Occurrence} from "../../utils/occurrence";
-import {Entity} from "../../utils/term";
+import {Entity} from "../../utils/entity";
 let assert = require("chai").assert;
 
 describe("Occurrence", () => {
