+++ src/utils/entity.ts
